GH-1119 Move impl into cpp file
